+++ trust_list/src/trust_list.c
@@ -48,7 +48,7 @@ vs_tl_init_storage(void) {
 
 /******************************************************************************/
 int
-vs_tl_save_part(vs_tl_element_info_t *element_info, const uint8_t *in_data, size_t data_sz) {
+vs_tl_save_part(vs_tl_element_info_t *element_info, const uint8_t *in_data, uint16_t data_sz) {
     if (NULL == element_info || NULL == in_data || element_info->id <= VS_TL_ELEMENT_MIN ||
         element_info->id >= VS_TL_ELEMENT_MAX) {
         return VS_TL_ERROR_PARAMS;
@@ -90,7 +90,7 @@ vs_tl_save_part(vs_tl_element_info_t *element_info, const uint8_t *in_data, size
 
 /******************************************************************************/
 int
-vs_tl_load_part(vs_tl_element_info_t *element_info, uint8_t *out_data, size_t buf_sz, size_t *out_sz) {
+vs_tl_load_part(vs_tl_element_info_t *element_info, uint8_t *out_data, uint16_t buf_sz, uint16_t *out_sz) {
     if (NULL == element_info || NULL == out_data || NULL == out_sz || element_info->id <= VS_TL_ELEMENT_MIN ||
         element_info->id >= VS_TL_ELEMENT_MAX) {
         return VS_TL_ERROR_PARAMS;
